refactor: Migrating Model file to TypeScript
